requirements.txt
langchain
langchain_core
langchain_community
langchain_google_genai
langchain_ollama
pypdf
langchain-google-genai
langchain-google-vertexai
google-generativeai
chromadb
beautifulsoup4
tiktoken
python-dotenv
langchain_chroma
fastapi
uvicorn
unstructured
gunicorn
pinecone
langchain_pinecone